+++ worksheet/Worksheet.js
@@ -4,7 +4,6 @@ var RelationshipManager = require("./RelationshipManager");
 /**
  * This module represents an excel worksheet in its basic form - no tables, charts, etc. Its purpose is
  * to hold data, the data's link to how it should be styled, and any links to other outside resources.
- * @module Excel/Worksheet
  */
 var Worksheet = /** @class */ (function () {
     /**
@@ -23,20 +22,13 @@ var Worksheet = /** @class */ (function () {
         this._orientation = null;
         this._margin = null;
         this.sharedStrings = { strings: {} };
-        // A two dimensional array of objects with custom XML attributes to add this worksheet's cells
-        // for example an object { style: 12b } at index [1][2] would add a {@code style="12b"} attribute to cell 'C2'
         this.customCellAttributes = [];
-        // A one dimensional array with the same purpose as the custom cell attributes array except the custom
-        // attributes are applied to the worksheet's rows
         this.customRowAttributes = [];
-        // The ID and settings for pageMargins and pageSetup
         this._printerSettings;
-        // An array of attributes to apply to the 'pageMargins' element of the spreadsheet
         this.pageMargins = null;
-        // An array of attributes (only "ref" Ex: "A1:D1" for now) to apply to the autoFilter element of the spreadsheet
         this.autoFilter = null;
-        // An array of attributes to apply to the 'pageSetup' element of the spreadsheet
         this.pageSetup = null;
+        this.dataValidations = null;
         var cfg = (config != null ? config : {});
         this.name = cfg.name;
         this.id = Util._uniqueId("Worksheet");
@@ -82,17 +74,23 @@ var Worksheet = /** @class */ (function () {
         this._drawings.push(table);
         this.relations.addRelation(table, "drawingRelationship");
     };
-    Worksheet.prototype.addPagePrintSetup = function (pageSetup, pageMargins, autoFilter) {
+    //public addPagePrintSetup() {
+    //    this._printerSettings = { id: Util.uniqueId('PrinterSettings') };
+    //    this.relations.addRelation(this._printerSettings, 'printerSettings');
+    //}
+    Worksheet.prototype.setAutoFilter = function (autoFilter) {
+        this.autoFilter = autoFilter;
+    };
+    Worksheet.prototype.setDataValidations = function (dataValidations) {
+        this.dataValidations = dataValidations;
+    };
+    Worksheet.prototype.setPageSetupAndMargins = function (pageSetup, pageMargins) {
         this.pageSetup = pageSetup;
         this.pageMargins = pageMargins;
-        this.autoFilter = autoFilter;
-        //this._printerSettings = { id: _.uniqueId('PrinterSettings') };
-        //this.relations.addRelation(this._printerSettings, 'printerSettings');
     };
     /** Expects an array length of three.
-     * @see Excel/Worksheet compilePageDetailPiece
+     * @see compilePageDetailPiece()
      * @see <a href='/cookbook/addingHeadersAndFooters.html'>Adding headers and footers to a worksheet</a>
-     *
      * @param headers [left, center, right]
      */
     Worksheet.prototype.setHeader = function (headers) {
@@ -102,9 +100,8 @@ var Worksheet = /** @class */ (function () {
         this._headers = headers;
     };
     /** Expects an array length of three.
-     * @see Excel/Worksheet compilePageDetailPiece
+     * @see compilePageDetailPiece()
      * @see <a href='/cookbook/addingHeadersAndFooters.html'>Adding headers and footers to a worksheet</a>
-     *
      * @param footers [left, center, right]
      */
     Worksheet.prototype.setFooter = function (footers) {
@@ -125,9 +122,7 @@ var Worksheet = /** @class */ (function () {
             "&R", this.compilePageDetailPiece(pieces[2] || "")
         ].join('');
     };
-    /** Turns instructions on page header/footer details into something
-     * usable by Excel.
-     *
+    /** Turns instructions on page header/footer details into something usable by Excel.
      * @param piece
      * @returns string | reduce
      */
@@ -185,8 +180,6 @@ var Worksheet = /** @class */ (function () {
     /** This creates some nodes ahead of time, which cuts down on generation time due to
      * most cell definitions being essentially the same, but having multiple nodes that need
      * to be created. Cloning takes less time than creation.
-     *
-     * @private
      * @param doc XmlDom
      * @returns
      */
@@ -213,30 +206,29 @@ var Worksheet = /** @class */ (function () {
     };
     /** Runs through the XML document and grabs all of the strings that will
      * be sent to the 'shared strings' document.
-     *
-     * @returns
+     * @returns list of cell 'value' strings with 'metadata.type' equal to 'text'
      */
     Worksheet.prototype.collectSharedStrings = function () {
         var data = this.data;
         var maxX = 0;
         var strings = {};
-        for (var row = 0, l = data.length; row < l; row++) {
-            var dataRow = data[row];
-            var cellCount = dataRow.length;
-            maxX = cellCount > maxX ? cellCount : maxX;
-            for (var c = 0; c < cellCount; c++) {
-                var cellValue = dataRow[c];
-                if (typeof dataRow[c] == "object") {
-                    cellValue = dataRow[c].value;
+        for (var r = 0, dLen = data.length; r < dLen; r++) {
+            var row = data[r];
+            var rLen = row.length;
+            maxX = rLen > maxX ? rLen : maxX;
+            for (var c = 0; c < rLen; c++) {
+                var cellValue = row[c];
+                if (typeof row[c] === "object") {
+                    cellValue = row[c].value;
                 }
-                var metadata = dataRow[c].metadata || {};
+                var metadata = row[c].metadata || {};
                 if (!metadata.type) {
-                    if (typeof cellValue == "number") {
+                    if (typeof cellValue === "number") {
                         metadata.type = "number";
                     }
                 }
                 if (metadata.type == "text" || !metadata.type) {
-                    if (typeof strings[cellValue] == "undefined") {
+                    if (typeof strings[cellValue] === "undefined") {
                         strings[cellValue] = true;
                     }
                 }
@@ -259,21 +251,21 @@ var Worksheet = /** @class */ (function () {
         var sheetData = Util.createElement(doc, "sheetData");
         var cellCache = this._buildCache(doc);
         var sharedStrs = this.sharedStrings;
-        for (var row = 0, l = data.length; row < l; row++) {
-            var dataRow = data[row];
-            var cellCount = dataRow.length;
-            maxX = cellCount > maxX ? cellCount : maxX;
+        for (var r = 0, dLen = data.length; r < dLen; r++) {
+            var row = data[r];
+            var rLen = row.length;
+            maxX = rLen > maxX ? rLen : maxX;
             var rowNode = doc.createElement("row");
-            for (var c = 0; c < cellCount; c++) {
+            for (var c = 0; c < rLen; c++) {
                 columns[c] = columns[c] || {};
-                var cellValue = dataRow[c];
-                if (cellValue != null && typeof cellValue == "object") {
+                var cellValue = row[c];
+                if (cellValue != null && typeof cellValue === "object") {
                     cellValue = cellValue.value;
                 }
                 //fix undefined or null value
-                var metadata = dataRow[c] ? (dataRow[c].metadata || {}) : {};
+                var metadata = row[c] ? (row[c].metadata || {}) : {};
                 if (!metadata.type) {
-                    if (typeof cellValue == "number") {
+                    if (typeof cellValue === "number") {
                         metadata.type = "number";
                     }
                     // Allows for empty cells in switch statement below
@@ -309,20 +301,20 @@ var Worksheet = /** @class */ (function () {
                 if (metadata.style) {
                     cell.setAttribute("s", metadata.style);
                 }
-                cell.setAttribute("r", Util.positionToLetterRef(c + 1, row + 1));
+                cell.setAttribute("r", Util.positionToLetterRef(c + 1, r + 1));
                 // add any additional custom attributes to this cell's XML element
-                if (row < customCellAttributes.length && customCellAttributes[row] != null && c < customCellAttributes[row].length) {
-                    var attribs = customCellAttributes[row][c];
+                if (r < customCellAttributes.length && customCellAttributes[r] != null && c < customCellAttributes[r].length) {
+                    var attribs = customCellAttributes[r][c];
                     for (var attrib in attribs) {
                         cell.setAttribute(attrib, attribs[attrib]);
                     }
                 }
                 rowNode.appendChild(cell);
             }
-            rowNode.setAttribute("r", row + 1);
+            rowNode.setAttribute("r", r + 1);
             // add any additional custom attributes to this row's XML element
-            if (row < customRowAttributes.length) {
-                var rowAttribs = customRowAttributes[row];
+            if (r < customRowAttributes.length) {
+                var rowAttribs = customRowAttributes[r];
                 for (var attrib in rowAttribs) {
                     rowNode.setAttribute(attrib, rowAttribs[attrib]);
                 }
@@ -353,7 +345,7 @@ var Worksheet = /** @class */ (function () {
         if (this._tables.length > 0) {
             var tables = doc.createElement("tableParts");
             tables.setAttribute("count", this._tables.length);
-            for (var i = 0, l = this._tables.length; i < l; i++) {
+            for (var i = 0, tLen = this._tables.length; i < tLen; i++) {
                 var table = doc.createElement("tablePart");
                 table.setAttribute("r:id", this.relations.getRelationshipId(this._tables[i]));
                 tables.appendChild(table);
@@ -362,7 +354,7 @@ var Worksheet = /** @class */ (function () {
         }
         if (this.mergedCells.length > 0) {
             var mergeCells = doc.createElement("mergeCells");
-            for (var i = 0, l = this.mergedCells.length; i < l; i++) {
+            for (var i = 0, mcLen = this.mergedCells.length; i < mcLen; i++) {
                 var mergeCell = doc.createElement("mergeCell");
                 mergeCell.setAttribute("ref", this.mergedCells[i][0] + ':' + this.mergedCells[i][1]);
                 mergeCells.appendChild(mergeCell);
@@ -393,10 +385,41 @@ var Worksheet = /** @class */ (function () {
             }
             worksheet.appendChild(pageSetupEl);
         }
-        for (var i = 0, l = this._drawings.length; i < l; i++) {
+        for (var i = 0, dLen = this._drawings.length; i < dLen; i++) {
             var drawing = doc.createElement("drawing");
             drawing.setAttribute("r:id", this.relations.getRelationshipId(this._drawings[i]));
             worksheet.appendChild(drawing);
+        }
+        // Add extLst element if there are custom dataValidations attributes
+        if (this.dataValidations) {
+            var extLstEl = doc.createElement("extLst");
+            var extChildEl = doc.createElement("ext");
+            extChildEl.setAttribute("xmlns:x14", "http://schemas.microsoft.com/office/spreadsheetml/2009/9/main");
+            extChildEl.setAttribute("uri", "{CCE6A557-97BC-4b89-ADB6-D9C93CAAB3DF}");
+            var dataValidationsEl = doc.createElement("x14:dataValidations");
+            dataValidationsEl.setAttribute("xmlns:xm", "http://schemas.microsoft.com/office/excel/2006/main");
+            dataValidationsEl.setAttribute("count", this.dataValidations.length.toString());
+            for (var i = 0; i < this.dataValidations.length; i++) {
+                var dataValadation = this.dataValidations[i];
+                var dataValidationEl = doc.createElement("x14:dataValidation");
+                dataValidationEl.setAttribute("showErrorMessage", "1");
+                dataValidationEl.setAttribute("showInputMessage", "1");
+                dataValidationEl.setAttribute("allowBlank", "1");
+                dataValidationEl.setAttribute("type", "list");
+                var formulaEl = doc.createElement("x14:formula1");
+                var xmfEl = doc.createElement("xm:f");
+                // Build formula to specify range where data set lies. Ex: SheetName!$A$5:$A$10
+                xmfEl.appendChild(doc.createTextNode(dataValadation.dataSheetName + "!$" + dataValadation.column + "$" + dataValadation.topRow + ":$" + dataValadation.column + "$" + dataValadation.bottomRow));
+                var sqrefEl = doc.createElement("xm:sqref");
+                sqrefEl.appendChild(doc.createTextNode(dataValadation.sqref));
+                formulaEl.appendChild(xmfEl);
+                dataValidationEl.appendChild(formulaEl);
+                dataValidationEl.appendChild(sqrefEl);
+                dataValidationsEl.appendChild(dataValidationEl);
+            }
+            extChildEl.appendChild(dataValidationsEl);
+            extLstEl.appendChild(extChildEl);
+            worksheet.appendChild(extLstEl);
         }
         return doc;
     };
@@ -433,7 +456,6 @@ var Worksheet = /** @class */ (function () {
         return cols;
     };
     /** Sets the page settings on a worksheet node.
-     *
      * @param doc XmlDom
      * @param worksheet XmlDom.XMLNode
      */
@@ -476,21 +498,18 @@ var Worksheet = /** @class */ (function () {
     };
     /** Can be one of 'portrait' or 'landscape'.
      * http://www.schemacentral.com/sc/ooxml/t-ssml_ST_Orientation.html
-     *
      * @param orientation
      */
     Worksheet.prototype.setPageOrientation = function (orientation) {
         this._orientation = orientation;
     };
     /** Expects an array of column definitions. Each column definition needs to have a width assigned to it.
-     *
      * @param columns
      */
     Worksheet.prototype.setColumns = function (columns) {
         this.columns = columns;
     };
     /** Expects an array of data to be translated into cells.
-     *
      * @param data Two dimensional array - [ [A1, A2], [B1, B2] ]
      * @see <a href='/cookbook/addingDataToAWorksheet.html'>Adding data to a worksheet</a>
      */
@@ -498,7 +517,6 @@ var Worksheet = /** @class */ (function () {
         this.data = data;
     };
     /** Merge cells in given range
-     *
      * @param cell1 - A1, A2...
      * @param cell2 - A2, A3...
      */